requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
librouteros==3.2.0
Flask
PyYAML
apscheduler